aider: Updated get_data function to assign 'NO DATA' for missing values in Screening.py
# Aider chat conversation:

USER: if the data isn't there i just want it to input NO DATA for a cell

ASSISTANT: To implement this change, we need to modify the `get_data` function in the `Screening.py` file. We will add a condition to check if the data is not available and if so, we will assign 'NO DATA' to the respective cell.

Here is the *SEARCH/REPLACE* block:

Screening.py
```python
<<<<<<< SEARCH
        info_all[agent].loc[paper_num, ['Paper Number',"Title", "Abstract"]] = [paper_num, title, abstract]
        if not bad_answer:
            info_all[agent].loc[paper_num, col_decision] = final_decision
            info_all[agent].loc[paper_num, col_initial] = initial_decision
            info_all[agent].loc[paper_num, col_thoughts] = thoughts
=======
        info_all[agent].loc[paper_num, ['Paper Number',"Title", "Abstract"]] = [paper_num, title, abstract] if title and abstract else 'NO DATA'
        if not bad_answer:
            info_all[agent].loc[paper_num, col_decision] = final_decision if final_decision else 'NO DATA'
            info_all[agent].loc[paper_num, col_initial] = initial_decision if initial_decision else 'NO DATA'
            info_all[agent].loc[paper_num, col_thoughts] = thoughts if thoughts else 'NO DATA'
>>>>>>> REPLACE
```
This change will ensure that if the data is not available, 'NO DATA' will be assigned to the respective cell.